+++ packages/contact-center/store/src/storeEventsWrapper.ts
@@ -16,9 +16,11 @@ import {
   ContactServiceQueue,
   Profile,
   AgentLoginProfile,
+  ERROR_TRIGGERING_IDLE_CODES,
 } from './store.types';
 import Store from './store';
 import {runInAction} from 'mobx';
+import {isIncomingTask} from './task-utils';
 
 class StoreWrapper implements IStoreWrapper {
   store: IStore;
@@ -50,7 +52,7 @@ class StoreWrapper implements IStoreWrapper {
   }
   get idleCodes() {
     return this.store.idleCodes.filter((code) => {
-      return code.name === 'RONA' || !code.isSystem;
+      return Object.values(ERROR_TRIGGERING_IDLE_CODES).includes(code.name) || !code.isSystem;
     });
   }
   get agentId() {
@@ -202,8 +204,8 @@ class StoreWrapper implements IStoreWrapper {
   };
 
   setCurrentTask = (task: ITask | null, isClicked: boolean = false): void => {
-    // Don't assign the task as current task if the interaction state is 'new' or 'consult'
-    if (task?.data.interaction.state === 'new' || task?.data.interaction.state === 'consult') return;
+    // Don't assign the task as current task is incoming
+    if (isIncomingTask(task)) return;
 
     runInAction(() => {
       // Determine if the new task is the same as the current task